requirements.txt
streamlit
requests
beautifulsoup4
rank_bm25
langchain
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
huggingface_hub
openai
python-dotenv
wikipedia
